+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Changes:
 
 - Add `flakeModules.default` and `modules.flake.default` to expose flake module too.
+- The `cspell.flake.app` option, which generates `apps.<system>.makeCspellConfig`, has been deprecated and replaced with the `cspell.packages.cspell-json` option, which generates `packages.<system>.cspell-json`.
 
 ## v1.0.0